pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "metadata-scrubber"
version = "0.1.0"
description = "CLI tool to scrub metadata from common file types"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Kaan" }]
keywords = ["metadata", "privacy", "scrub", "exif", "pdf", "docx"]
classifiers = [
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.11",
]

dependencies = [
  "typer>=0.12,<1.0",
  "rich>=13,<14",
  "pillow>=10,<12",
  "pypdf>=4,<6",
  "defusedxml>=0.7,<1.0",
]

[project.optional-dependencies]
audio = ["mutagen>=1.47,<2.0"]

[project.scripts]
metadata-scrubber = "metadata_scrubber.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
